requirements.txt
aiogram==2.24.1
aiosqlite
python-dotenv
pandas
openpyxl